+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Changed
+- Select minimal widget is now scrollable.
+
 [4.5.12] - 2016-03-28
 ---------------------
 ##### Changed