+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.5.13] - 2016-04-01
 ---------------------
 ##### Changed
 - Select minimal widget is now scrollable and won't stretch form.